+++ django/applications/catmaid/static/libs/catmaid/connectivity_matrix.js
@@ -33,6 +33,11 @@
     // dimensional array. Each [row][col] entry contains two values:
     // outgoing from rows to columns, incoming from columns to rows.
     this.connectivityMatrix = [];
+
+    // A set of filter rules to apply to the handled connectors
+    this.filterRules = [];
+    // Filter rules can optionally be disabled
+    this.applyFilterRules = true;
   };
 
   /**
@@ -42,13 +47,88 @@
   ConnectivityMatrix.prototype.refresh = function() {
     // Return a promise that is fullfilled, if the table is ready
     var self = this;
+    var with_locations = self.filterRules && self.applyFilterRules;
+
     return CATMAID.fetch(project.id + '/skeleton/connectivity_matrix', 'POST', {
           'rows': self.rowSkeletonIDs,
-          'columns': self.colSkeletonIDs
+          'columns': self.colSkeletonIDs,
+          'with_locations': with_locations
         })
+      .then(function(json) {
+        return self.filterResults(json);
+      })
       .then(function(json) {
         self.setConnectivityMatrixFromData(json);
       });
+  };
+
+  ConnectivityMatrix.prototype.filterResults = function(data) {
+    var hasResults = data && !CATMAID.tools.isEmpty(data);
+    if (this.filterRules.length > 0 && this.applyFilterRules && hasResults) {
+      // Collect connector models from input
+      var connectors = new Map();
+      for (var sourceSkeletonId in data) {
+        var targets = data[sourceSkeletonId];
+        for (var targetSkeletonId in targets) {
+          var links = targets[targetSkeletonId];
+          if (!links.locations) {
+            throw new CATMAID.ValueError("Could not find location information in connectivity matrix");
+          }
+          for (var connectorId in links.locations) {
+            var connector  = links.locations[connectorId];
+            connectorId = parseInt(connectorId, 10);
+            if (!connectors.has(connectorId)) {
+              var x = connector.pos[0];
+              var y = connector.pos[1];
+              var z = connector.pos[2];
+              connectors.set(connectorId,
+                  new CATMAID.ConnectorModel(connectorId, x, y, z));
+            }
+          }
+        }
+      }
+      // Filter connectors
+      var filter = new CATMAID.NodeFilter(this.filterRules, connectors);
+      return filter.execute()
+        .then(function(filtered) {
+          if (filtered.nodes.size === 0) {
+            CATMAID.warn("No connectors left after filter application");
+            return Promise.resolve([]);
+          }
+          // Filter links
+          let allowedNodes = new Set(Object.keys(filtered.nodes).map(function(n) {
+            return parseInt(n, 10);
+          }));
+          for (var sourceSkeletonId in data) {
+            var targets = data[sourceSkeletonId];
+            for (var targetSkeletonId in targets) {
+              var links = targets[targetSkeletonId];
+              if (!links.locations) {
+                throw new CATMAID.ValueError("Could not find location information in connectivity matrix");
+              }
+              for (var connectorId in links.locations) {
+                connectorId = parseInt(connectorId, 10);
+                if (!allowedNodes.has(connectorId)) {
+                  var loc = links.locations[connectorId];
+                  links.count = links.count - loc.count;
+                  delete links.locations[connectorId];
+                }
+              }
+
+              if (links.count === 0) {
+                delete targets[targetSkeletonId];
+              }
+            }
+
+            if (CATMAID.tools.isEmpty(targets)) {
+              delete data[sourceSkeletonId];
+            }
+          }
+
+          return Promise.resolve(data);
+        });
+    }
+    return Promise.resolve(data);
   };
 
   /**
@@ -80,7 +160,7 @@
     for (var i=0; i<this.rowSkeletonIDs.length; ++i) {
       m[i] = new Array(this.colSkeletonIDs.length);
       for (var j=0; j<this.colSkeletonIDs.length; ++j) {
-        m[i][j] = 0;
+        m[i][j] = { 'count': 0, 'data': {} };
       }
     }
 
@@ -104,7 +184,18 @@
         var rowSourceIndex = rowIndexCache[sourceSkid];
         var colPartnerIndex = colIndexCache[partnerSkid];
         if (rowSourceIndex !== undefined && colPartnerIndex !== undefined) {
-          m[rowSourceIndex][colPartnerIndex] = partners[partnerSkid];
+          var count = partners[partnerSkid];
+          if (typeof(count) === 'number') {
+            m[rowSourceIndex][colPartnerIndex] = {
+              'count': count,
+              'data': {}
+            };
+          } else {
+            m[rowSourceIndex][colPartnerIndex] = {
+              'count': count.count,
+              'data': count.locations
+            };
+          }
         }
       }
     }